이메일로 임시 비밀번호 받는 기능을 구현했습니다.
테스트페이지에서 가입할때 작성할 이메일을 쓰고 임시 비밀번호 전송을 누르면
서블릿에서 임시비밀번호 난수생성 -> 해당난수 이메일로 발송 -> 임시비밀번호를 특정키 ( playddit + 유저 아이디 + 유저비밀번호) 를 이용한 AES256 기반으로 암호화 ( 관리자도 비밀번호를 파싱 할 수 없도록) 해서 DB에 저장 하도록 구현했습니다.
src.util 패키지에 이를 위해서 CryptoUtil 과 EmailUtil 을 각각 만들어 추가했습니다.
