+++ playddit/res/mapper/users.xml
@@ -38,8 +38,8 @@
 		 where user_nickname = #user_nickname#
 	</select>
 	
-	<!-- 임시 비밀번호 설정 -->
-	<update id="sendTempPass2" parameterClass="map">
+	<!-- 비밀번호 변경 -->
+	<update id="setNewPass" parameterClass="map">
 		update users
 		set user_password = #password#
 		where user_id = #email#
