--- playddit/src/users/main/signUp/VerifyEmail.java
@@ -1,105 +1,0 @@
-package users.main.signUp;
-
-import java.util.Properties;
-
-import javax.mail.Message;
-import javax.mail.MessagingException;
-import javax.mail.PasswordAuthentication;
-import javax.mail.Session;
-import javax.mail.Transport;
-import javax.mail.internet.AddressException;
-import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeMessage;
-
-public class VerifyEmail {
-	
-	public static void sendEmail(String id, String pw) {
-		String recipient = id;
-		String code = pw;
-		
-		final String user = "ddit302@gmail.com";
-		final String password = "ddit.or.kr.302";
-		Properties prop = new Properties();
-		prop.put("mail.smtp.host", "smtp.gmail.com");
-		prop.put("mail.smtp.port", 465);
-		prop.put("mail.smtp.auth", "true");
-		prop.put("mail.smtp.ssl.enable", "true");
-		prop.put("mail.smtp.ssl.trust", "smtp.gmail.com");
-		
-		Session session = Session.getDefaultInstance(prop, new javax.mail.Authenticator() {
-			protected PasswordAuthentication getPasswordAuthentication() {
-				return new PasswordAuthentication(user, password);
-			}
-		});
-		
-		// 4. Message 클래스의 객체를 사용하여 수신자와 내용, 제목의 메시지를 작성한다.
-		// 5. Transport 클래스를 사용하여 작성한 메세지를 전달한다.
-		
-		MimeMessage message = new MimeMessage(session);
-		try {
-			message.setFrom(new InternetAddress(user));
-			
-			message.addRecipient(Message.RecipientType.TO, new InternetAddress(recipient));
-			message.setSubject("PLAYDDIT 임시 비밀번호 안내");
-			message.setText("임시비밀번호를 안내해 드립니다. 괄호 안의 값만 입력해주세요 ["+code+"]");
-			Transport.send(message);	// send message
-			
-		} catch (AddressException e) {
-			e.printStackTrace();
-		} catch (MessagingException e) {
-			e.printStackTrace();
-		}
-	}
-	
-	public static void main(String[] args) {
-		String recipient = "psh40963@naver.com";
-		String code = "abc";
-		
-		// 1. 발신자의 메일 계정과 비밀번호 설정
-		final String user = "ddit302@gmail.com";
-		final String password = "ddit.or.kr.302";
-		
-		// 2. Property에 SMTP 서버 정보 설정
-		Properties prop = new Properties();
-		prop.put("mail.smtp.host", "smtp.gmail.com");
-		prop.put("mail.smtp.port", 465);
-		prop.put("mail.smtp.auth", "true");
-		prop.put("mail.smtp.ssl.enable", "true");
-		prop.put("mail.smtp.ssl.trust", "smtp.gmail.com");
-		
-		// 3. SMTP 서버정보와 사용자 정보를 기반으로 Session 클래스의 인스턴스 생성
-		Session session = Session.getDefaultInstance(prop, new javax.mail.Authenticator() {
-			protected PasswordAuthentication getPasswordAuthentication() {
-				return new PasswordAuthentication(user, password);
-			}
-		});
-		
-		// 4. Message 클래스의 객체를 사용하여 수신자와 내용, 제목의 메시지를 작성한다.
-		// 5. Transport 클래스를 사용하여 작성한 메세지를 전달한다.
-		
-		MimeMessage message = new MimeMessage(session);
-		try {
-			message.setFrom(new InternetAddress(user));
-			
-			// 수신자 메일 주소
-			message.addRecipient(Message.RecipientType.TO, new InternetAddress(recipient));
-			
-			// Subject
-			message.setSubject("PLAYDDIT verification code");
-			
-			// Text
-			message.setText("Welcome to playddit. your code is ["+code+"]");
-			
-			Transport.send(message);	// send message
-			System.out.println("이메일 전송을 성공적으로 완료했습니다.");
-			
-			
-		} catch (AddressException e) {
-			e.printStackTrace();
-		} catch (MessagingException e) {
-			e.printStackTrace();
-		}
-
-	}
-
-}
